requirements.txt
Flask==2.2.5
Flask-Cors==5.0.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
requests==2.32.3
pillow==9.4.0
tensorflow==2.17.0
